+++ lib/src/core/interface/app_component.dart
